+++ js/data.js
@@ -1,0 +1,31 @@
+const INdescription=[
+  'Передо мной интересная фотография ',
+  'Я думаю, что на ней изображён поход',
+  'Давайте рассмотрим изображение внимательнее',
+  'На переднем плане фотографии, где и находится центр событий, мы видим двух девочек',
+  'Они радостны'
+];
+
+const INname=[
+  'Кирилл',
+  'Матвей',
+  'Михаил',
+  'Вера',
+  'Данила',
+  'Богдан',
+  'Марат',
+  'Марк',
+  'Анфиса',
+  'Денис'
+];
+
+const INmessage=['Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+
+export{INdescription};
+export{INname};
+export{INmessage};
